third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/flow-content-0/slot-element-tabbable.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] slot element with display: block should be focusable
  assert_equals: slot is now focused expected Element node <slot tabindex="1" style="display: block"></slot> but got null
[FAIL] slot element with default style should be focusable
  assert_equals: slot is now focused expected Element node <slot tabindex="1" style=""></slot> but got null
Harness: the test ran to completion.
